fix: move return on error decorators under one dir
