+++ server/decorators/__init__.py
@@ -1,2 +1,1 @@
-from .return_on_404_decorator import return_on_404
-from .return_on_error_decorator import return_on_error
+from .return_on_error import return_on_error, return_on_404
